style(fmt): make clippy happy
